EEH-2513 - First batch of code restructuration
